project clean u

+++ src/main/kotlin/space/yaroslav/familybot/executors/command/AskWorldInitialExecutor.kt
@@ -112,7 +112,7 @@ class AskWorldInitialExecutor(
     ) {
         val questionWithIds = question.copy(
             id = questionId.await(),
-            messageId = result.messageId + chat.id.toLong()
+            messageId = result.messageId + chat.id
         )
         askWorldRepository.addQuestionDeliver(questionWithIds, chat)
     }
